util: handle QUOTE_0_TERMINATED|QUOTE_EXPECT_TRAILING_0 in printstr_ex
There is a number of places in kernel where the last character
of a string-typed structure field is not fetched and/or overwritten with
'\0'.  So far, it was difficult to signify this in strace output done
via printstr_ex; this commit adds logic that decreases size passed
to print_quoted_string in order to signify that, while there
is an additional non-NUL character in the field, it will be ignored
by the kernel.

* src/util.c (printstr_ex): Decrease size before print_quoted_string
call if both QUOTE_0_TERMINATED and QUOTE_EXPECT_TRAILING_0 set.